I need checkout

+++ app/Http/Controllers/OrderController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use App\Models\Order;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -9,33 +10,37 @@ use App\Http\Requests\Order as OrderRequests;
 use App\Events\OrderUpdated;
 use Carbon\Carbon;
 
-class OrderController extends Controller{
+class OrderController extends Controller
+{
     protected $order;
 
-    public function __construct(Order $order){
+    public function __construct(Order $order)
+    {
         $this->order = $order;
     }
     /**
      * Muestra lista de ordenes
      * @return Json api rest
      */
-    public function index(Request $request){
-        $orderby = $request->orderby? $request->orderby : "created_at";
-        $order = $request->order=="desc"? "desc" : "asc";
+    public function index(Request $request)
+    {
+        $orderby = $request->orderby ? $request->orderby : "created_at";
+        $order = $request->order == "desc" ? "desc" : "asc";
         $page = $request->itemsPage ? $request->itemsPage : 20;
-        $rol = Auth::user()->rol;
-        
-        //Validation for branchs to admins
-        if(!$rol){
-            if(!isset($request->branch)) $branch = Auth::user()->branch_id;
-            else {
-                if($request->branch === "all") $branch = null;
-                else $branch = $request->branch;
+        $currentUser = Auth::user();
+        $branchUser = $currentUser->branch_id;
+        $branch = $branchUser;
+
+        // If branches var is not present, use the same branch of user
+        // only admin can see all branches
+        if (isset($request->branch)) {
+            if ($request->branch === "all") {
+                $branch = null;
+            } else {
+                $branch = $request->branch;
             }
-        }else {
-            $branch = Auth::user()->branch_id;
         }
-        
+
         $orderdb = $this->order
             ->withRelation()
             ->Estado($request->status)
@@ -54,21 +59,26 @@ class OrderController extends Controller{
      * @param  $request de body en Json
      * @return Json api rest
      */
-    public function store(OrderRequests $request){
-        $request['user_id']= Auth::user()->id;
-        $order = $this->order->create($request->all());
-        $rol = Auth::user()->rol;
-        //Validation for branchs to admins
-        if(!$rol){
-            if(!isset($request['branch_id'])) $request['branch_id'] = Auth::user()->branch_id; 
-        }else {
-            $request['branch_id'] = Auth::user()->branch_id; 
-        } 
+    public function store(OrderRequests $request)
+    {
+        $currentUser = Auth::user();
+        $request['user_id'] = $currentUser->id;
+        $request['branch_id'] = $currentUser->branch_id;
+        $request['status'] = 0;
+        $rolUser = $currentUser->rol;
 
-        if(isset($request->items)){
-            $order['items'] = $this->getItemsRequest($request->items);
-            if(count($order['items'])) event(new OrderUpdated($order, false));
+        //Only admin can save in differents branches
+        if (!$rolUser) {
+            if (isset($request->branch_id)) $request['branch_id'] = $request->branch_id;
         }
+
+        $order = $this->order->create($request->all());
+
+        if (isset($request->items)) {
+            $order['items'] = $this->getItemsRequest($request->items, $request['branch_id']);
+            if (count($order['items'])) event(new OrderUpdated($order, false));
+        }
+
 
         return new OrderResources($order);
     }
@@ -78,7 +88,8 @@ class OrderController extends Controller{
      * @param  $order identificador de la orden
      * @return Json api rest
      */
-    public function show(Order $order){
+    public function show(Order $order)
+    {
         $order->withRelation();
         return new OrderResources($order);
     }
@@ -89,18 +100,28 @@ class OrderController extends Controller{
      * @param  $order identificador de la orden a actualizar
      * @return Json api rest
      */
-    public function update(Request $request, Order $order){
-        $request['user_id']=$order->user_id;
+    public function update(Request $request, Order $order)
+    {
+        $currentUser = Auth::user();
+        $request['updated_id'] = $currentUser->id;
         $udStatus = $order->status != $request->status ? true : false;
-        
-        $order->update( $request->all() );
-        if(isset($request->items)){
-            $order['items'] = $this->getItemsRequest($request->items);
-            
-            if(count($order['items'])) event(new OrderUpdated($order, $udStatus));
+        $rolUser = $currentUser->rol;
+        //Only admin can modify branches
+        if (isset($request->branch_id) && $rolUser) {
+            unset($request['branch_id']);
         }
-    
-        return New OrderResources($order);
+
+        if ($order) {
+            $order->update($request->all());
+
+            if (isset($request->items)) {
+                $order['items'] = $this->getItemsRequest($request->items);
+
+                if (count($order['items'])) event(new OrderUpdated($order, $udStatus));
+            }
+        }
+
+        return new OrderResources($order);
     }
 
     /**
@@ -108,28 +129,46 @@ class OrderController extends Controller{
      * @param  $order identificador de la orden
      * @return null 404
      */
-    public function destroy($id){
+    public function destroy($id)
+    {
         $order = $this->order::where('id', $id)
-                ->with('nota')
-                ->first();
+            ->with('nota')
+            ->first();
+        $enUso = 0;
 
-        $enUso = count($order->nota);
+        if ($order) {
+            $enUso = count($order->nota ?? []);
 
-        if($enUso){
-            $order->deleted_at = Carbon::now();
-            $order->updated_id = Auth::user()->id;
-            $order->save();
-        } else {
-            $order->delete();
+            if ($enUso) {
+                $order->deleted_at = Carbon::now();
+                $order->updated_id = Auth::user()->id;
+                $order->save();
+            } else {
+                $order->delete();
+            }
         }
-        
-        //$order->delete();
+
         return response()->json(null, 204);
     }
 
-    private function getItemsRequest($items){
-        if($items) return is_string($items) ? json_decode($items, true) : $items;
+    private function getItemsRequest($items, $branch_id = null)
+    {
+        if ($items) {
+            $itemsArray = is_string($items) ? json_decode($items, true) : $items;
 
-        return null;
+            if (is_array($itemsArray)) {
+                if ($branch_id) {
+                    foreach ($itemsArray as $key => $item) {
+                        if (!isset($item['branch_id'])) {
+                            $itemsArray[$key]['branch_id'] = $branch_id;
+                        }
+                    }
+                }
+
+                return $itemsArray;
+            }
+        }
+
+        return [];
     }
 }
